Splitting pokemonUtils in two parts
